tornado_handlers: reduce duplicated code
- create a common base class TornadoRequestHandlerBase
- split out a common method get_generated_db_data_from_log
