+++ pylintrc
@@ -215,7 +215,7 @@ spelling-store-unknown-words=no
 [SIMILARITIES]
 
 # Minimum lines number of a similarity.
-min-similarity-lines=4
+min-similarity-lines=7
 
 # Ignore comments when computing similarities.
 ignore-comments=yes
